requirements.txt
langchain>=0.1.0
langchain-community>=0.0.12
langchain-core>=0.1.0
langchain-openai>=0.0.8
langchain-anthropic>=0.1.0
langchain-text-splitters>=0.0.1

# 文档处理
pypdf>=3.17.0
faiss-cpu>=1.7.4

# 图像处理
pillow>=9.5.0

# HTTP 请求
requests>=2.31.0
playwright>=1.40.0

# 其他依赖
python-dotenv>=1.0.0
tqdm>=4.65.0
openai>=1.0.0
anthropic>=0.39.0
sniffio>=1.3.1
distro>=1.9.0

pymupdf
pillow
duckduckgo-search